introduces TBird::Namer
This class is responsibile for creating a unique name for the file and
follows this pattern:

    "#{identifier}/#{token}_#{version}.#{extension}

The identifier is user settable and defaults to a SHA1 hash of the
original filename.

The token is also user settable and defaults to a generated UUID.

The version defaults to 'original' and corresponds to the name of the
version given to the image after being transformed by TBird::Processor
(e.g. thumbnail).