introduces TBird::Processor
This class is responsible for applying MiniMagick transformations to an
image to create multiple versions of the same file. It has basic
commands for creating a square thumbnail and resizing an image (using
width, height, width x height, percentage or any other geomotry commands understood by
MiniMagick) as well as providing an interface for passing a block into
MiniMagick's `combine_options` method to allow for more advanced
tranformations.