introduces TBird::Transmitter
This class is responsible for uploading the file(s) to S3:

    t = TBird::Transmitter.new('filename', data, metadata)
    t.transmit!
    t.url

It is oriented for streaming uploads (to support large files) and allows
the user to pass in arbitrary key/value's in the metadata hash that will
be written with the file to S3.

Two metadata keys are special, :access (defaults to :public_read)
controls the ACL used on S3 and :content_type (defaults to
'binary/octet-stream').