introduces TBird::Uploader
This is the main class of this library and will likely be the only entry
point for most users. Right now it handles parsing out the filename,
filedata and content type from an uploaded file (e.g. pass it your
params[:file] after the form has been posted), but will soon also be
responsible for using the other classes, Namer, Processor and
Transmitter, to create versions of the file and upload them to S3 with
their new names.

The return value will be a hash of the URLs to the file(s) on S3, keyed
by their version name. More soon.